* master: (1056 commits)
  Applied SP policy
  Fixed a bug that made the maintenance popup appear 2 times. Fixed bug that mad the game crash if the force upgrade was not sent as an http error
  Mark as override ToString method
  ToString method added to obfuscated classes
  Passing constant values as parameters
  Implicit conversions added to derived types. Unit tests added as well.
  Unit tests added for Obfuscation classes
  Fixes in subtraction operators
  Mask computation fix. Avoiding thread unsafe code
  Ignore big unscaled delta times in game time
  build.settings was using wrong unity version 5.3 (instead of 5.5) and wrong ndk (r10d instead of r10e)
  updated PlayServicesResolver to the latest available (v1.2.31.0)
  added play services resolver as an external module (version 1.2.12)
  fixing some warnings from External libraries
  update unity test tools from 1.5.8 to 1.5.9 (fixes some unity 5.5 warnings)
  buildsets fixes
  disabled shadow casting/receiving, motion vectors and probes in spine mesh renderer
  Memory obfuscation classes
  Added deploy to prod ServerTask Added icon to this task
  Fixed ScenesData scriptable object empty when repo checkout (save assets call was required).
  ...

+++ Project/Assets/Sparta/SocialPoint/Social/IPlayerData.cs
@@ -18,6 +18,6 @@
         /// <summary>
         /// Player Level
         /// </summary>
-        long Level { get; }
+        int Level { get; }
     }
 }
